test(linux): share bundle behavior coverage with Windows
